+++ crates/core/src/lib.rs
@@ -1,6 +1,5 @@
 #![feature(more_qualified_paths)]
 pub mod client;
-pub mod kakarot;
+pub mod contracts;
 pub mod mock;
 pub mod models;
-pub mod test_utils;
